Changed the format of the provided information each turn and at the beginning of the game

+++ src/game/turns/config.rs
@@ -1,14 +1,15 @@
+use bevy::prelude::Resource;
 use serde::Deserialize;
 
-#[derive(Debug, Deserialize)]
+#[derive(Debug, Deserialize, Copy, Clone, Resource)]
 #[serde(default)]
 pub struct TurnConfig {
     #[serde(rename = "timeout")]
-    pub turn_time: u64,
+    pub turn_time: u32,
     #[serde(rename = "wait")]
     pub wait_for_all: bool,
     #[serde(rename = "turns")]
-    pub max_turns: u64,
+    pub max_turns: u32,
 }
 impl Default for TurnConfig {
     fn default() -> Self {
